requirements.txt
aiohttp
Pillow
Shapely
gdal
tqdm
